Merge in COMPSCI3307_F2021/group31 from MainGame to master

* commit 'a9a6c4eee5234f395edb4b2350b7348428ff1d3d':
  Typing feature works and is improved

+++ GameClass.h
@@ -12,7 +12,7 @@ class GameClass {
         TextObject *gameText;
         int missedChars;
         time_t startTime, endTime;
-        int progressPercent;
+        float progressPercent;
         //StatsController statistics Access;
         int currentIndex;
